requirements.txt
Flask
requests
python-dotenv
streamlit
bs4